[#2522] authorized keys owner is verified
The user method on the provider always returned what the resource should
be, not what it actually was, so it always seemed to be insync to
puppet.

Also cleaned up some cruft on the perms that did different things
depending on whether a user was specified on the resource.  This isn't
necessary since a user is required on the resource.

Paired with: Jesse Wolfe
Signed-off-by: Matt Robinson <matt@puppetlabs.com>