[#4090] Rename mode (:master, :agent, :user) to run_mode
Mode is a terribly overused word.  Files use it, puppetdoc uses it, and
certs use it, and those are just the places that I happened to
stumble upon.  It makes reading code very confusing and finding things
in code difficult.  I know namespacing allows us to reuse words for
method and variable names, but that doesn't mean we should.