Minimal fix for #4297, with notes for follow-up
In retrospect it appears that the fix for #4270 was incomplete and somewhat off
target.  This patch fixes the one demonstrably incorrect part (the namespace)
and adds a comment outlining what remains to be done to clean up the code;
these additional changes, while needed for maintanability, are inappropriate
for a quick turnaround crucial bug fix release such as 2.6.1, at which this
patch is targeted.