(#4763) Don't call a method that was removed in Rails 3 activerecord
Calling this method caused storeconfigs not to run.

ActiveRecord::Base.allow_concurrency was deprecated in Rails 2.2.  We
support activerecord 2.1 and higher, so we still need to call this
method for 2.1.  I factored out the code that determines our
activerecord version to a method in util so that the code was easier to
read and test.