(#5660) Changed wrong description for protocol
I contrast to the description you cannot use the same title for
different resources. Thats what title_patterns are for: To use one title
to specify all key_attributes. So if one wants to specify a resource
with two protocols, the easiest way to this is to just write

  port { 'telnet:tcp': number => 23, ensure => present }
  port { 'telnet:udp': number => 23, ensure => present }

The titlepattern will both set name ("telnet") and protocol (:tcp or
:udp)