(#5660) porttype use :tcp as default protocol
The default value for a keyattribute like protocol in port.rb does only
work when the titlepattern does not set it (even if the titlepattern
will set the parameter to nil).

Now the first titlepattern searches for tcp or udp (e.g.
telnet:udp) and if this titlepattern does not match the resourcetitle
only :name will be set in the parameter hash. If protocol is not
explicitly set puppet will finally take the default value into account
(which is :tcp)