maint: finish transition of application help to return strings.
Jesse made a change, in e1191f33defcaffec5900c7122a89ca75d3a9673, to
transition from printing and exiting in the help method up to returning the
help data to the caller.  This was part of eliminating rdoc usage from the
display of help to the user.

The cert application was missed, and still used the legacy "print and exit"
model; this cleans that up so it matches the rest of the code.

Paired-With: Matt Robinson <matt@puppetlabs.com>