(#6962) Better argument checking for help.
We used to blink and miss the fact that we failed to load an action or face in
the past; now we test for that, and fail with a clear error message when the
user asks for something we can't deliver.

Additionally, fix a couple of tests that were silently broken because they
passed the wrong arguments, but still got some output.

Paired-With: Matt Robinson <matt@puppetlabs.com>