maint: Remove utf-8 character in comment
Daniel's emacs inserts a differnt kind of dash.  It doesn't seem to
break anything, but the `# -*- coding: utf-8 -*-` comment doesn't seem
to matter either and isn't even the format that Ruby 1.9 uses for
specifying encoding.  To avoid problems it's probably best to just not put
special characters in code unless necessary.

This can easily just be squashed in somewhere too.