(#6442) Be able to start agents --listen without namespaceauth.conf
The namespaceauth.conf isn't being used since we switched from XMLRPC to
Rest.  There's still a lot of cleanup of the code that deals with
authorization from the namespaceauth.conf and XMLRPC in general, but
that can be address later.  For now just being able to start puppet
agent in listen mode without a useless file will be a big win.  You
still need the auth.conf file though since that is used.

Reviewed-by: Max Martin <max@puppetlabs.com>