(#11552) Remove TransObject and TransBucket
TransObject and TransBucket were no longer used except as an intermediate form
when converting between Puppet::Type::WhateverType and Puppet::Resource. This removes
TransBucket and TransObject entirely and rewrites the conversion (to_resource)
so the intermediate to_trans call is unneeded. There was custom logic in
Puppet::Type::File#to_trans that was moved to a to_resource method.

Backward compatibility with 0.24.x has been broken for some time (0.25.5
is the last release I've seen that works running a master with 0.24.x
agents), but we removed Network::Handler::Master which depended on
TransObject. This is the beginning of removing XMLRPC code
(ticket #8296), which is what in theory provided the 0.24.x backward
compatibility.

Paired with Patrick Carlisle <patrick@puppetlabs.com>