(#8296) Remove XMLRPC code
The XMLRPC code was left in place to ensure backward compatibility with
Puppet 0.24.x clients, but 0.24.x clients haven't been able to talk to
masters since 0.25.6 due to other bugs and incompatibilities.
Therefore, the XMLRPC code has been dead code for a long time and
removing it makes the codebase a lot easier to navigate.

TODO:
* should also be able to remove the sslcertificates code, because only
  the ssl/ directory should be used by the REST code.