Fix message when changing to purged from absent
    $ puppet apply -e "package { 'ack-grep' : ensure => purged }"
    notice: /Stage[main]//Package[ack-grep]/ensure: created

Getting a message that something was created when it's purged is wrong.  You
would get this both in error in #3707 when the ensure state wasn't really being
changed since absent and purged are really the same thing for rpm, but you
would also get this in apt based systems if you purged a package that had
previously just been absent.

It's not ideal to have logic for an ensure state that is only relevant to a few
types in property.rb, but it still seemed like the best place to fix this.

TODO:
tests
think if there's a better place for the logic to live than property.rb