(#3707) Fix yum to know when something is purged
The distinction between purged and absent in yum is actually not necessary like
it is for apt.  However, the logic for insync? in the package type considers a
package to be absent if purged, but not purged if absent.  This is why when the
rpm query method ended up showing the current state as absent, Puppet would try
to purge again even though it didn't need to.

The fix that was done earlier by James Turnbull for this issue to have rpm
query show uninstalled pacakges as purged actually is basically correct, it
just needed some logic in the install method for the yum provider so that it
could tell when bogus package names are attempted and error instead of
pretending like the install happend successfully.

TODO before pull request:
TESTS!
Ask someone who uses Redhat based systems if this fix makes sense