Update module and environment with functionality for the module tool
Being able to figure out what other modules depend on a module is useful
for resolving contraints when installing and preventing uninstalls of
modules that are still required.  And being able to tell if a module has
local changes is useful to prevent accidentally losing those changes
when upgrading or uninstalling a module.